* Add fall_back_on_butt animation to `animation.yml` * Change name of `StarDustHandler.kt` to `ShootingStarHandler.kt` to better describe the classes purpose * Moved various methods from `ShootingStar.kts` to `ShootingStarHandler.kt` * Add player animation when moved due to being under a crashing star
